19/04/2025 group api

+++ lib/sourcing/OnBoarding/house_visit_form.dart
@@ -174,8 +174,9 @@ class _HouseVisitFormState extends State<HouseVisitForm> {
 
       //double lat1 = 28.541909899980464;
       //double lon1 = 77.23837911402565;
-      double lat1 = widget.GroupData.latitude ?? 0.0;
-      double lon1 = widget.GroupData.longitude ?? 0.0;
+      double lat1 = (widget.GroupData.latitude ?? 0.0) as double;
+      double lon1 = (widget.GroupData.longitude ?? 0.0) as double;
+
 
 
       double distance = calculateDistance(lat1, lon1, _latitude, _longitude);
